+++ cloud-proxy-app/Program.cs
@@ -1,71 +1,42 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Microsoft.Extensions.Configuration;
+﻿using Microsoft.Extensions.DependencyInjection;
+using System;
 
 namespace Glasswall.IcapServer.CloudProxyApp
 {
     class Program
     {
-        const string InputConfigurationKey = "input";
-        const string OutputConfigurationKey = "output";
-        const string ConfigurationConfigurationKey = "configuration";
+        static ServiceProvider _serviceProvider;
 
         static int Main(string[] args)
         {
-            var configuration = GetConfiguration(args);
-            if (!CheckConfigurationIsValid(configuration))
-            {
-                return (int)ReturnOutcome.GW_ERROR;
-            }
-
-            return CopyFile(configuration[InputConfigurationKey], configuration[OutputConfigurationKey]);
-        }
-
-        private static int CopyFile(string sourceFilepath, string destinationFilepath)
-        {
             try
             {
-                var destinationFolder = System.IO.Path.GetDirectoryName(destinationFilepath);
-                System.IO.Directory.CreateDirectory(destinationFolder);
-                System.IO.File.Copy(sourceFilepath, destinationFilepath, true);
-                return (int)ReturnOutcome.GW_REBUILT;
+                RegisterServices();
+                IServiceScope serviceScope = _serviceProvider.CreateScope();
+                return serviceScope.ServiceProvider.GetRequiredService<CloudProxyApplication>().Run(args);
             }
-            catch(Exception ex)
+            finally
             {
-                Console.WriteLine($"Error Processing 'input' {sourceFilepath}, {ex.Message}");
-                return (int)ReturnOutcome.GW_ERROR;
+                DisposeServices();
             }
         }
 
-        static bool CheckConfigurationIsValid(IConfiguration configuration)
+        private static void DisposeServices()
         {
-            var configurationErrors = new List<string>();
+            if (_serviceProvider == null)
+                return;
 
-            if (string.IsNullOrEmpty(configuration[InputConfigurationKey]))
-                configurationErrors.Add($"'{InputConfigurationKey}' is missing from command line");
-
-            if (string.IsNullOrEmpty(configuration[OutputConfigurationKey]))
-                configurationErrors.Add($"'{OutputConfigurationKey}' is missing from command line");
-
-            if (configurationErrors.Any())
+            if (_serviceProvider is IDisposable disposable)
             {
-                Console.WriteLine($"Error Processing Command {Environment.NewLine} \t{string.Join(',', configurationErrors)}");
+                disposable.Dispose();
             }
-
-            return !configurationErrors.Any();
         }
 
-        static IConfiguration GetConfiguration(string[] args)
+        private static void RegisterServices()
         {
-            var switchMappings = new Dictionary<string, string>()
-            {
-                { "-i", InputConfigurationKey },
-                { "-o", OutputConfigurationKey },
-                { "-c", ConfigurationConfigurationKey },
-            };
-            var builder = new ConfigurationBuilder().AddCommandLine(args, switchMappings);
-            return builder.Build();
+            var services = new ServiceCollection();
+            services.AddSingleton<CloudProxyApplication>();
+            _serviceProvider = services.BuildServiceProvider(true);
         }
     }
 }
